+++ fuc/cli/vcf_rename.py
@@ -37,7 +37,7 @@ def create_parser(subparsers):
     )
     parser.add_argument(
         'vcf',
-        help='VCF file (zipped or unzipped).'
+        help='VCF file (compressed or uncompressed).'
     )
     parser.add_argument(
         'names',